Updated AUT custom classes files

+++ fse2019/splittypie/src/main/java/po/event/components/AddEditTransactionComponent.java
@@ -44,7 +44,7 @@ public class AddEditTransactionComponent extends BasePageObject implements PageC
         this.type(By.xpath("//input[contains(@class, \"transaction-name\")]"), transactionName);
     }
 
-    public void typeAmount(int amount){
+    public void typeAmount(double amount){
         this.type(By.xpath("//input[contains(@class, \"transaction-amount\")]"), String.valueOf(amount));
     }
