+++ fse2019/splittypie/src/main/java/po_apogen/EditTransactionComponent.java
@@ -51,7 +51,7 @@ public class EditTransactionComponent extends BasePageObject implements PageComp
 		this.type(By.xpath("//input[contains(@class, \"transaction-name\")]"), transactionName);
 	}
 
-	public void typeAmount(int amount){
+	public void typeAmount(double amount){
 		this.type(By.xpath("//input[contains(@class, \"transaction-amount\")]"), String.valueOf(amount));
 	}
